# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ShopNow Electronics API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products with optional filters
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: featured
          in: query
          schema:
            type: boolean
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
...
  /products/search:
    get:
      operationId: searchProducts
      tags: [products]
      summary: Search products with text query and filters
      parameters:
        - name: q
          in: query
          schema:
            type: string
          description: Text search query (matches name, brand, category)
        - name: category
          in: query
          schema:
            type: string
        - name: department
          in: query
          schema:
            type: string
          description: "Filter by department, e.g. Gaming"
        - name: componentType
          in: query
          schema:
            type: string
          description: "Filter by component type, e.g. Processor"
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: inStock
          in: query
          schema:
            type: boolean
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [relevance, price_asc, price_desc, rating, newest]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Search results with pagination
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductSearchResult"
...
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
...
components:
  schemas:
...
    Product:
      type: object
      required:
        - id
        - name
        - brand
        - price
        - category
        - imageUrl
        - rating
        - reviewCount
        - inStock
      properties:
        id:
          type: integer
        name:
          type: string
        brand:
          type: string
        price:
          type: number
        originalPrice:
          type: ["number", "null"]
        discountPct:
          type: ["integer", "null"]
        category:
          type: string
        department:
          type: ["string", "null"]
        componentType:
          type: ["string", "null"]
        externalId:
          type: ["string", "null"]
        sourceUrl:
          type: ["string", "null"]
        imageUrl:
          type: string
        images:
          type: ["string", "null"]
          description: "JSON array of image URLs"
        rating:
          type: number
        reviewCount:
          type: integer
        inStock:
          type: boolean
        stockCount:
          type: ["integer", "null"]
        specs:
          type: ["string", "null"]
          description: "JSON object of product specifications"
        isFeatured:
          type: boolean
        isDeal:
          type: boolean
...
    ProductSearchResult:
      type: object
      required:
        - products
        - total
        - page
        - totalPages
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
...